Restructured the project a bit splitting logic to follow the MVC design pattern.
