+++ src/bot.py
@@ -13,15 +13,14 @@ from src.cogs.leave import Leave
 from src.cogs.play import Play
 from src.cogs.test import TestCog
 from src.models.bot_config import get_bot, get_bot_token, get_guild_id
-from src.threads import get_thread_warden
-
-
-log.getLogger(__name__)  # Set same logging parameters as main.py.
+from src.util.threads import get_thread_warden
+from src.models.stream_thread import main_stream_thread
 
 
 ##########################################################################
-######################   CORE APPLICATION LOGIC   ########################
+#########################  APPLICATION SETUP   ###########################
 ##########################################################################
+
 
 async def load_cogs(bot: commands.Bot, guild_id: discord.Object) -> None:
     """
@@ -32,12 +31,13 @@ async def load_cogs(bot: commands.Bot, guild_id: discord.Object) -> None:
     try:
         # await bot.load_extension(f"src.cogs.test")
 
+        # * {guild_id} is user-defined.
         await bot.add_cog(Join(bot), guilds=[guild_id])
         await bot.add_cog(Leave(bot), guilds=[guild_id])
         await bot.add_cog(Play(bot), guilds=[guild_id])
         await bot.add_cog(TestCog(bot), guilds=[guild_id])
 
-        log.debug(f"Successfully loaded the cogs!")
+        log.debug(f"Succesffully loaded the cogs!")
 
     except TypeError as t_err:
         log(f"Type error, failed to load the cogs: %s", t_err)
@@ -54,20 +54,19 @@ async def load_cogs(bot: commands.Bot, guild_id: discord.Object) -> None:
     return
 
 
-async def run_discord_bot() -> None:
+async def main_bot() -> None:
     """
     Function acts as the entry point for the Discord bot and is used to
         instantiate (start) the bot instance.
     """
     try:
         bot = get_bot()
-        guild_id = get_guild_id()
         bot_token = get_bot_token()
+        guild_id = get_guild_id()
+
         async with bot:
-            # * {guild_id} is user-defined and read from the .env file.
-            #! await load_cogs(bot, guild_id)
-            
-            # * {bot_token} is user-defined and read from the .env file.
+            await load_cogs(bot, guild_id)
+            # * {bot_token} is user-defined.
             await bot.start(bot_token)
 
     except discord.DiscordException as d_err:
@@ -80,8 +79,9 @@ async def run_discord_bot() -> None:
 
 
 ##########################################################################
-#########################   EVENT LISTENERS   ############################
+##########################   EVENT LISTENERS   ###########################
 ##########################################################################
+
 
 @get_bot().event
 async def on_ready() -> None:
@@ -91,21 +91,21 @@ async def on_ready() -> None:
     """
     log.info("Groovester started Successfully!")
     print("Groovester started Successfully!")
-    
-    # Sync slash commands on Discord server.
-    await bot.tree.sync(guild=guild_id)
 
-    # Start various helper threads.
-    play_songs_in_discord_audio_thread = Thread(
-        target=run_discord_audio_thread, args=()
-    )
+    # Run worker threads.
+    main_stream_thread()
 
-    try:
-        play_songs_in_discord_audio_thread.start()
+    # # Start various helper threads.
+    # play_songs_in_discord_audio_thread = Thread(
+    #     target=run_discord_audio_thread, args=()
+    # )
 
-    except Exception as err:
-        log.error(f"General Exception, on_ready failed to spawn child thread: {err}")
-        #! TODO: Kill process when this exception is thrown.
+    # try:
+    #     play_songs_in_discord_audio_thread.start()
+
+    # except Exception as err:
+    #     log.error(f"General Exception, on_ready failed to spawn child thread: {err}")
+    #     #! TODO: Kill process when this exception is thrown.
 
     #! TODO: Remove because Discord.py doesn't support slash commands with
     #!  cog structure.
@@ -120,29 +120,3 @@ async def on_ready() -> None:
     #     )
 
     return
-
-def run_discord_audio_thread() -> None:
-    """
-    Function used to start a new thread dedicated to preparing and
-        streaming songs via voice channels.
-    """
-    #! TODO: The reason this function is needed, is because on_ready
-    #!  uses Thread class, which can't run asynchronous functions.
-    #!  See if asyncio can resolve this and remove this function.
-    asyncio.run(get_thread_warden().prepare_discord_audio())
-
-    return
-
-
-# def get_bot() -> commands.Bot:
-#     """
-#     Temp text.
-#     """
-#     return bot
-
-
-# def get_guild_id() -> discord.Object:
-#     """
-#     Temp text.
-#     """
-#     return guild_id
