+++ src/cogs/join.py
@@ -4,8 +4,8 @@ from time import sleep
 import discord
 from discord.ext import commands
 
-from src.helpers import get_voice_client, is_connected, set_voice_client
-
+from src.util.helpers import is_connected
+from src.models.stream_thread import set_voice_client
 
 
 log.getLogger(__name__)
@@ -51,7 +51,7 @@ class Join(commands.Cog):
                 voice_channel = message_author.voice.channel
                 set_voice_client(await voice_channel.connect())
                 # print(type(get_voice_client()))
-                
+
                 log.debug(
                     f"!join successfully connected to the voice channel: {voice_channel.name}"
                 )
