+++ src/cogs/leave.py
@@ -4,7 +4,8 @@ import random
 import discord
 from discord.ext import commands
 
-from src.helpers import is_connected, set_voice_client
+from src.util.helpers import is_connected
+from src.models.stream_thread import set_voice_client
 
 
 log.getLogger(__name__)
