+++ test/helpers.py
@@ -4,24 +4,22 @@ from pathlib import Path
 # from pytube import YouTube
 from pytubefix import YouTube
 
-from src.helpers import *
+from src.util.helpers import *
 
 
 def test_downloading_youtube_video():
     """
-    Function to test that the `pytubefix` library can download YouTube 
-        videos. Tests the functionality of various file system helper 
+    Function to test that the `pytubefix` library can download YouTube
+        videos. Tests the functionality of various file system helper
         functions too.
     """
     assert setup_media_directory()
 
     # Download a random YouTube video.
     youtube_obj = YouTube("https://youtu.be/QC8iQqtG0hg?si=zJXhXDfwGs7rWn74")
-    audio_stream = youtube_obj.streams.get_audio_only(
-        subtype="mp4"
-    )
+    audio_stream = youtube_obj.streams.get_audio_only(subtype="mp4")
     path_to_file = Path(audio_stream.download(filename=f"{youtube_obj.video_id}.mp4"))
-    
+
     # Verify the file actually exists on the file system, then remove it.
     assert os.path.exists(path_to_file)
     assert remove_media_file(path_to_file)
